+++ components/src/lib/types/api.d.ts
@@ -152,6 +152,8 @@ export type JobStatus = 'submitted' | 'done' | 'failed';
 
 export type ImagesStatus = {
 	status: JobStatus;
+} | {
+	status: JobStatus;
 	num_submitted_images: number;
 	num_done_images: number;
 	num_failed_images: number;