+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Added first and last task index to job info modal (\#818);
+* Displayed hourglass icon on submitted tasks that have not run yet (\#818);
+
 # 1.19.1
 
 * Used separate environment variables (`PUBLIC_FRACTAL_VIZARR_VIEWER_URL` and `PUBLIC_FRACTAL_DATA_URL`) to configure fractal-data URL and vizarr viewer URL (\#816);